pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssl-certificate-monitor"
version = "1.0.0"
description = "Monitor SSL/TLS certificate expiry across multiple domains with alerting and reporting"
authors = [{ name = "Pratham Biyani", email = "prathambiyani85@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
keywords = ["ssl", "tls", "certificate", "monitoring", "security"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "jinja2>=3.1",
    "certifi>=2024.0",
]

[project.scripts]
ssl-monitor = "ssl_monitor.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=ssl_monitor --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools]
packages = ["ssl_monitor"]
